move BrowseResponseReceived to BrowseService and ClearSearchHistory to SearchTabHelper
