+++ Seeker/Browse/BrowseService.cs
@@ -1,4 +1,4 @@
-using Android.Content;
+﻿using Android.Content;
 using Android.Views;
 using Android.Widget;
 using Common;
@@ -18,6 +18,13 @@ namespace Seeker.Browse
 {
     public static class BrowseService
     {
+        public static event EventHandler<BrowseResponseEvent> BrowseResponseReceived;
+
+        public static void OnBrowseResponseReceived(BrowseResponse origBR, TreeNode<Directory> rootTree, string fromUsername, string startingLocation)
+        {
+            BrowseResponseReceived?.Invoke(null, new BrowseResponseEvent(origBR, rootTree, fromUsername, startingLocation));
+        }
+
         public static void GetFolderContentsAPI(string username, string dirname, bool isLegacy, Action<Task<IReadOnlyCollection<Directory>>> continueWithAction)
         {
             if (!PreferencesState.CurrentlyLoggedIn)
@@ -123,7 +130,7 @@ namespace Seeker.Browse
                 var tree = CreateTree(br.Result, false, null, null, username, out errorString);
                 if (tree != null)
                 {
-                    SeekerState.OnBrowseResponseReceived(br.Result, tree, username, atLocation);
+                    BrowseService.OnBrowseResponseReceived(br.Result, tree, username, atLocation);
                 }
 
                 SeekerState.ActiveActivityRef.RunOnUiThread(() =>
